backend/requirements_linux.txt
altgraph==0.17.3
amqp==5.1.1
antlr4-python3-runtime==4.10
async-timeout==4.0.2
bidict==0.22.0
build==0.8.0
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
eventlet==0.33.1
Flask==2.1.2
Flask-SocketIO==5.2.0
greenlet==1.1.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
kombu==5.2.4
MarkupSafe==2.1.1
mouse==0.7.1
MouseInfo==0.1.3
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pep517==0.13.0
Pillow==9.2.0
PyAutoGUI==0.9.53
pycairo==1.21.0
PyGetWindow==0.0.9
PyGObject==3.42.2
pyinstaller==5.5
pyinstaller-hooks-contrib==2022.10
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.9
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.1
python3-xlib==0.15
pytweening==1.0.4
six==1.16.0
tomli==2.0.1
typer==0.4.1
vine==5.0.0
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
